updated fact table to use inner join instead of left join

+++ sql_scripts/5. staging_shop_dim.sql
@@ -1,12 +1,11 @@
 --create if not exist
-DROP TABLE computer_std.dbo.dim_shop;
+IF OBJECT_ID('computer_std.dbo.dim_shop', 'U') IS NOT NULL
+    DROP TABLE computer_std.dbo.dim_shop;
 CREATE TABLE computer_std.dbo.dim_shop (
     Shop_ID INT IDENTITY(1,1) PRIMARY KEY,
     [Shop Name] VARCHAR(250) NULL,
     [Shop Age] INT NULL
 );
-
-
 --insert data into the table
 INSERT INTO computer_std.dbo.dim_shop 
 	([Shop Name], [Shop Age])
@@ -16,16 +15,3 @@ FROM computer_std.dbo.pc_data;
 
 --select all data from the dim table
 SELECT * FROM computer_std.dbo.dim_shop;
-
-WITH ShopCleaner AS (
-    SELECT 
-        Shop_ID, 
-        [Shop Name],
-        ROW_NUMBER() OVER (
-            PARTITION BY [Shop Name] 
-            ORDER BY Shop_ID
-        ) AS RowNum
-    FROM computer_std.dbo.dim_shop
-)
-DELETE FROM ShopCleaner 
-WHERE RowNum > 1;
